Move some of misc1.c to indent.c.
My little contribution to #209.
