+++ src/edit.c
@@ -22,6 +22,7 @@
 #include "fileio.h"
 #include "fold.h"
 #include "getchar.h"
+#include "indent.h"
 #include "main.h"
 #include "mbyte.h"
 #include "memline.h"
