+++ src/misc1.h
@@ -1,12 +1,6 @@
 #ifndef NEOVIM_MISC1_H
 #define NEOVIM_MISC1_H
 /* misc1.c */
-int get_indent(void);
-int get_indent_lnum(linenr_T lnum);
-int get_indent_buf(buf_T *buf, linenr_T lnum);
-int get_indent_str(char_u *ptr, int ts);
-int set_indent(int size, int flags);
-int get_number_indent(linenr_T lnum);
 int open_line(int dir, int flags, int second_line_indent);
 int get_leader_len(char_u *line, char_u **flags, int backward,
                    int include_space);
